Compute probability distribution tables

Compute and save to disk probability distributions of random variables from extracted ALFRESCO outputs.

alf_dist(j, in_dir, out_dir, period, reps, project = NULL)

Arguments

j
integer, iterator, the sequence of unique geographic regions available among the extracted data.

in_dir
input directory where extracted data are located in .rds files.

out_dir
output directory where random variables distribution tables are saved as .rds files.

period
character, "historical" or "projected".

reps
integer vector, simulation replicates included in data extraction, e.g., 1:200. Uses all replicates if not provided.

project
character, inferred from in_dir if project = NULL. Use this for project with subprojects/treatment group Alfresco runs.

Details

This function estimates discrete probability distributions of random variables from extracted ALFRESCO model runs. The random variables available include fire size, burn area and fire frequency by vegetation class, as well as cover area and age for a vegetation class.

This function estimates probability distributions for outputs from run_alf_extraction.

Examples

# NOT RUN {
mclapply(1:n, alf_dist, in_dir, out_dir, period = period, mc.cores = mc.cores)
# }