Fire point probabilities

Estimate fire point probability from fire weights for point locations.

fire_probs(data, veg_labels, covariates = TRUE)

Arguments

data
a data frame output by fire_weights.

veg_labels
character, vector of vegetation labels. If missing, they are assumed from get_veg_labels("ak1km").

covariates
logical, if TRUE (default), then a wide data frame is returned that includes additional spatially aggregated covariates. See details.

Value

a data frame.

Details

This function estimates point probability of fire for spatial point locations from fire weights output by fire_weights. It generates a summary data frame that includes a column prop of a simple proportion of area burned and a column prob of estimated fire point probabilities based on the weights in data. covariates = TRUE returns additional variables only if at least the veg column is present in data. If not, there will be nothing additional to return.

This function returns a one-column data frame. It is called on a single point location just like fire_weights. It is generally used in a context where it is invoked multiple times applied to several point locations. See point_probs for the generic wrapper used to estimate fire point probability for multiple locations that returns a multi-row data frame.

Examples

# NOT RUN {
scar <- "FireScar_0_2004.tif"
veg <- "Veg_0_2004.tif"
age <- "Age_0_2004.tif"
coords <- data.frame(lon = -147.7164, lat = 64.8378)
fire_weights(r = scar, xy = coords, veg = veg, age = age)
fire_probs(x)
# }